services/orders-api/pyproject.toml
[tool.poetry]
name = "asrp-orders-api"
version = "0.1.0"
description = "Orders API for asRP (enterprise-grade, OIDC/JWT, RBAC)"
authors = ["asrp <dev@local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.30"
psycopg = {extras = ["binary"], version = "^3.2.0"}
alembic = "^1.13.2"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
python-json-logger = "^2.0.7"
PyJWT = "^2.9.0"
cryptography = "^43.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
ruff = "^0.5.0"
mypy = "^1.10.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
